Move and rename `updateSelectionByTabClick` to `mouse-event-listener.js`
